Created Wrapper component. All service provider logic has been moved to him.
