+++ src/components/Wrapper/index.js
@@ -1,0 +1,3 @@
+import Wrapper from './Wrapper';
+
+export default Wrapper;
