+++ src/pages/404.js
@@ -1,16 +1,19 @@
 import React from 'react';
 
+import Wrapper from '@components/Wrapper';
 import Layout from '@components/Layout';
 import Head from '@components/Head';
 import Title from '@scenes/404/components/Title';
 
 const NotFound = () => (
-  <Layout>
-    <Head>
-      <title>Not Found - Halo Lab Blog</title>
-    </Head>
-    <Title />
-  </Layout>
+  <Wrapper>
+    <Layout>
+      <Head>
+        <title>Not Found - Halo Lab Blog</title>
+      </Head>
+      <Title />
+    </Layout>
+  </Wrapper>
 );
 
 export default NotFound;
