+++ src/pages/blog.js
@@ -1,16 +1,19 @@
 import React from 'react';
 
+import Wrapper from '@components/Wrapper';
 import Layout from '@components/Layout';
 import Head from '@components/Head';
 
 const Blog = () => {
   return (
-    <Layout>
-      <Head>
-        <title>Blog - Halo Lab Blog</title>
-      </Head>
-      <div>Blog</div>
-    </Layout>
+    <Wrapper>
+      <Layout>
+        <Head>
+          <title>Blog - Halo Lab Blog</title>
+        </Head>
+        <div>Blog</div>
+      </Layout>
+    </Wrapper>
   );
 };
