+++ src/pages/contacts.js
@@ -1,16 +1,19 @@
 import React from 'react';
 
+import Wrapper from '@components/Wrapper';
 import Layout from '@components/Layout';
 import Head from '@components/Head';
 
 const Contacts = () => {
   return (
-    <Layout>
-      <Head>
-        <title>Contacts - Halo Lab Blog</title>
-      </Head>
-      <div>Contacts</div>
-    </Layout>
+    <Wrapper>
+      <Layout>
+        <Head>
+          <title>Contacts - Halo Lab Blog</title>
+        </Head>
+        <div>Contacts</div>
+      </Layout>
+    </Wrapper>
   );
 };
