+++ src/pages/portfolio.js
@@ -1,16 +1,19 @@
 import React from 'react';
 
+import Wrapper from '@components/Wrapper';
 import Layout from '@components/Layout';
 import Head from '@components/Head';
 
 const Portfolio = () => {
   return (
-    <Layout>
-      <Head>
-        <title>Portfolio - Halo Lab Blog</title>
-      </Head>
-      <div>Portfolio</div>
-    </Layout>
+    <Wrapper>
+      <Layout>
+        <Head>
+          <title>Portfolio - Halo Lab Blog</title>
+        </Head>
+        <div>Portfolio</div>
+      </Layout>
+    </Wrapper>
   );
 };
