+++ src/pages/services.js
@@ -1,5 +1,6 @@
 import React from 'react';
 
+import Wrapper from '@components/Wrapper';
 import Layout from '@components/Layout';
 import Head from '@components/Head';
 import MailUs from '@components/MailUs';
@@ -11,19 +12,21 @@ import Technologies from '@scenes/Services/components/Technologies';
 
 const Services = () => {
   return (
-    <Layout>
-      <Head>
-        <title>Services - Halo Lab Blog</title>
-      </Head>
-      <div className="pageWrapper">
-        <Design />
-        <Development />
-        <Flow />
-        <Industries />
-        <Technologies />
-        <MailUs />
-      </div>
-    </Layout>
+    <Wrapper>
+      <Layout>
+        <Head>
+          <title>Services - Halo Lab Blog</title>
+        </Head>
+        <div className="pageWrapper">
+          <Design />
+          <Development />
+          <Flow />
+          <Industries />
+          <Technologies />
+          <MailUs />
+        </div>
+      </Layout>
+    </Wrapper>
   );
 };
