+++ src/pages/thanks.js
@@ -1,16 +1,19 @@
 import React from 'react';
 
+import Wrapper from '@components/Wrapper';
 import Layout from '@components/Layout';
 import Head from '@components/Head';
 
 const Thanks = () => {
   return (
-    <Layout>
-      <Head>
-        <title>Thanks - Halo Lab Blog</title>
-      </Head>
-      <div>Thanks</div>
-    </Layout>
+    <Wrapper>
+      <Layout>
+        <Head>
+          <title>Thanks - Halo Lab Blog</title>
+        </Head>
+        <div>Thanks</div>
+      </Layout>
+    </Wrapper>
   );
 };
